feat: add modular helpers for interface stub generation and request processing
- Introduced `Emitter.Helpers.cs`, `Parser.Helpers.cs`, and `Parser.Request.Helpers.cs` to modularize and centralize helper functions for generated stubs and request handling logic.
- Moved query and header-handling logic to `RequestBuilderImplementation.QueryAndHeaders.Helpers.cs` for improved organization and maintainability.
- Refactored existing logic to leverage shared helpers, reducing redundancy and increasing code clarity.
