+++ src/InterfaceStubGenerator.Shared/Parser.Request.cs
@@ -9,10 +9,6 @@ namespace Refit.Generator;
 
 /// <summary>Parses candidate interfaces and methods into the models used to generate Refit stubs.</summary>
 /// <content>Request parsing helpers for the Refit source generator.</content>
-[SuppressMessage(
-    "Style",
-    "SST1202:Members should be ordered by accessibility",
-    Justification = "Internal helper seams are kept next to their production call sites to preserve parser readability.")]
 internal static partial class Parser
 {
     /// <summary>Parses the request metadata needed by generated request construction.</summary>
@@ -32,11 +28,7 @@ internal static partial class Parser
 
         var httpMethodAttribute = FindHttpMethodAttribute(
             methodSymbol,
-            context.HttpMethodBaseAttributeSymbol);
-        if (httpMethodAttribute is null)
-        {
-            return RequestModel.Empty;
-        }
+            context.HttpMethodBaseAttributeSymbol)!;
 
         var httpMethod = GetHttpMethodName(httpMethodAttribute.AttributeClass);
         var path = GetHttpPath(httpMethodAttribute);
@@ -65,25 +57,6 @@ internal static partial class Parser
             parameters);
     }
 
-    /// <summary>Finds the HTTP method attribute on a Refit method.</summary>
-    /// <param name="methodSymbol">The method to inspect.</param>
-    /// <param name="httpMethodAttribute">The Refit HTTP method base attribute symbol.</param>
-    /// <returns>The matching attribute, if any.</returns>
-    private static AttributeData? FindHttpMethodAttribute(
-        IMethodSymbol methodSymbol,
-        INamedTypeSymbol httpMethodAttribute)
-    {
-        foreach (var attributeData in methodSymbol.GetAttributes())
-        {
-            if (attributeData.AttributeClass?.InheritsFromOrEquals(httpMethodAttribute) == true)
-            {
-                return attributeData;
-            }
-        }
-
-        return null;
-    }
-
     /// <summary>Gets the HTTP method name represented by a Refit method attribute.</summary>
     /// <param name="attributeClass">The attribute type.</param>
     /// <returns>The HTTP method name, or an empty string for unsupported custom attributes.</returns>
@@ -110,61 +83,6 @@ internal static partial class Parser
         return arguments.Length > 0 && arguments[0].Value is string path
             ? path
             : string.Empty;
-    }
-
-    /// <summary>Determines whether the initial inline emitter can use the path as a literal URI.</summary>
-    /// <param name="path">The path to inspect.</param>
-    /// <returns><see langword="true"/> when the path is supported.</returns>
-    internal static bool IsConstantPathSupported(string path) =>
-        (path.Length == 0 || path[0] == '/')
-        && path.IndexOf('{') < 0
-        && path.IndexOf('}') < 0
-        && path.IndexOf('\r') < 0
-        && path.IndexOf('\n') < 0;
-
-    /// <summary>Normalizes constant inline paths to match the reflection request builder URI cleanup.</summary>
-    /// <param name="path">The source path from the HTTP method attribute.</param>
-    /// <returns>The normalized path used by generated request construction.</returns>
-    internal static string NormalizeConstantPathForInline(string path)
-    {
-        var fragmentIndex = path.IndexOf('#');
-        if (fragmentIndex >= 0)
-        {
-            path = path[..fragmentIndex];
-        }
-
-        var queryIndex = path.IndexOf('?');
-        if (queryIndex < 0)
-        {
-            return path;
-        }
-
-        var pathOnly = path[..queryIndex];
-        if (queryIndex == path.Length - 1)
-        {
-            return pathOnly;
-        }
-
-        var queryStart = queryIndex + 1;
-        var partStart = queryStart;
-        char[]? queryBuffer = null;
-        var queryLength = 0;
-
-        for (var i = queryStart; i <= path.Length; i++)
-        {
-            if (i < path.Length && path[i] != '&')
-            {
-                continue;
-            }
-
-            var partLength = i - partStart;
-            AppendNonEmptyQueryPart(path, queryStart, partStart, partLength, ref queryBuffer, ref queryLength);
-            partStart = i + 1;
-        }
-
-        return queryBuffer is null
-            ? pathOnly
-            : $"{pathOnly}?{new string(queryBuffer, 0, queryLength)}";
     }
 
     /// <summary>Appends a query segment unless its key is empty or whitespace.</summary>
@@ -204,30 +122,6 @@ internal static partial class Parser
 
         path.CopyTo(partStart, queryBuffer, queryLength, partLength);
         queryLength += partLength;
-    }
-
-    /// <summary>Determines whether a string slice is empty or all whitespace.</summary>
-    /// <param name="value">The string containing the slice.</param>
-    /// <param name="start">The slice start index.</param>
-    /// <param name="length">The slice length.</param>
-    /// <returns><see langword="true"/> if the slice is empty or all whitespace.</returns>
-    internal static bool IsWhiteSpace(string value, int start, int length)
-    {
-        if (length <= 0)
-        {
-            return true;
-        }
-
-        var end = start + length;
-        for (var i = start; i < end; i++)
-        {
-            if (!char.IsWhiteSpace(value[i]))
-            {
-                return false;
-            }
-        }
-
-        return true;
     }
 
     /// <summary>Determines whether a method carries request metadata the initial inline emitter does not handle.</summary>
@@ -308,47 +202,11 @@ internal static partial class Parser
                         AddStaticHeader(headers, header);
                     }
                 }
-
-                continue;
             }
 
-            if (argument.Value is string singleHeader)
-            {
-                AddStaticHeader(headers, singleHeader);
-            }
+            // HeadersAttribute is declared as params string[], so Roslyn always exposes
+            // values as an array typed constant for supported Refit metadata.
         }
-    }
-
-    /// <summary>Adds one static header to the final header list.</summary>
-    /// <param name="headers">The mutable header list.</param>
-    /// <param name="header">The raw header declaration.</param>
-    internal static void AddStaticHeader(List<HeaderModel> headers, string header)
-    {
-        if (string.IsNullOrWhiteSpace(header))
-        {
-            return;
-        }
-
-        var separator = header.IndexOf(':');
-        var name = separator >= 0
-            ? header[..separator].Trim()
-            : header.Trim();
-        var value = separator >= 0
-            ? header[(separator + 1)..].Trim()
-            : null;
-
-        for (var i = 0; i < headers.Count; i++)
-        {
-            if (!string.Equals(headers[i].Name, name, StringComparison.Ordinal))
-            {
-                continue;
-            }
-
-            headers[i] = new(name, value);
-            return;
-        }
-
-        headers.Add(new(name, value));
     }
 
     /// <summary>Parses request parameter bindings for the conservative initial inline path.</summary>
@@ -695,61 +553,6 @@ internal static partial class Parser
         return false;
     }
 
-    /// <summary>Tries to parse a body buffered constructor argument.</summary>
-    /// <param name="argument">The constructor argument.</param>
-    /// <param name="buffered">Receives the buffered value.</param>
-    /// <returns><see langword="true"/> when the argument is a boolean buffered argument.</returns>
-    internal static bool TryGetBodyBufferedValue(in TypedConstant argument, out bool buffered)
-    {
-        if (argument is
-            {
-                Type.SpecialType: SpecialType.System_Boolean,
-                Value: bool boolValue
-            })
-        {
-            buffered = boolValue;
-            return true;
-        }
-
-        buffered = false;
-        return false;
-    }
-
-    /// <summary>Gets the Refit body serialization enum member name for an underlying value.</summary>
-    /// <param name="value">The enum value.</param>
-    /// <returns>The enum member name.</returns>
-    internal static string GetBodySerializationMethodName(int value) =>
-        value switch
-        {
-            0 => "Default",
-            1 => "Json",
-            BodySerializationUrlEncoded => "UrlEncoded",
-            BodySerializationSerialized => "Serialized",
-            _ => string.Empty
-        };
-
-    /// <summary>Determines whether all body bindings are supported by the initial inline emitter.</summary>
-    /// <param name="parameters">The parsed request parameter models.</param>
-    /// <returns><see langword="true"/> when every body binding is supported.</returns>
-    internal static bool IsSupportedInlineBody(ImmutableEquatableArray<RequestParameterModel> parameters)
-    {
-        foreach (var parameter in parameters)
-        {
-            if (parameter.Kind != RequestParameterKind.Body)
-            {
-                continue;
-            }
-
-            if (parameter.BodySerializationMethod.Length == 0
-                || parameter.BodySerializationMethod == "UrlEncoded")
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
     /// <summary>Gets return-type details required by the shared generated request runner.</summary>
     /// <param name="methodSymbol">The Refit method symbol.</param>
     /// <returns>The parsed return type details.</returns>
@@ -791,25 +594,18 @@ internal static partial class Parser
     /// <param name="resultType">The method result type.</param>
     /// <param name="isApiResponse">Whether the result is an API response wrapper.</param>
     /// <returns>The deserialization target type.</returns>
-    private static string GetDeserializedResultTypeName(ITypeSymbol resultType, bool isApiResponse) =>
-        !isApiResponse
-            ? resultType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)
-            : resultType switch
-            {
-                INamedTypeSymbol { MetadataName: "IApiResponse" } => "global::System.Net.Http.HttpContent",
-                INamedTypeSymbol { TypeArguments.Length: 1 } namedType => namedType.TypeArguments[0]
-                    .ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat),
-                _ => resultType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat)
-            };
+    private static string GetDeserializedResultTypeName(ITypeSymbol resultType, bool isApiResponse)
+    {
+        if (!isApiResponse)
+        {
+            return resultType.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
+        }
 
-    /// <summary>Determines whether the shared runner should dispose the response.</summary>
-    /// <param name="deserializedResultType">The deserialized result type.</param>
-    /// <returns><see langword="true"/> when the runner owns response disposal.</returns>
-    internal static bool ShouldDisposeResponse(string deserializedResultType) =>
-        deserializedResultType is not
-            "global::System.Net.Http.HttpResponseMessage" and not
-            "global::System.Net.Http.HttpContent" and not
-            "global::System.IO.Stream";
+        var namedType = (INamedTypeSymbol)resultType;
+        return namedType.MetadataName == "IApiResponse"
+            ? "global::System.Net.Http.HttpContent"
+            : namedType.TypeArguments[0].ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
+    }
 
     /// <summary>Parsed body attribute data.</summary>
     /// <param name="SerializationMethod">The body serialization method name.</param>
