+++ src/Refit/PushStreamContent.cs
@@ -147,10 +147,6 @@ namespace System.Net.Http
         /// <remarks>
         /// https://github.com/ASP-NET-MVC/aspnetwebstack/blob/5118a14040b13f95bf778d1fc4522eb4ea2eef18/src/Common/TaskHelpers.cs#L65.
         /// </remarks>
-        [SuppressMessage(
-            "Design",
-            "SST1436:Add members to the type or remove it",
-            Justification = "Intentional empty placeholder struct for the void TaskCompletionSource pattern.")]
         private readonly struct AsyncVoid : IEquatable<AsyncVoid>
         {
             /// <inheritdoc/>
