+++ src/Refit/RequestBuilderImplementation.QueryAndHeaders.cs
@@ -4,7 +4,6 @@
 
 using System.Collections;
 using System.Diagnostics.CodeAnalysis;
-using System.Globalization;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Web;
@@ -159,113 +158,6 @@ namespace Refit
             }
 
             return vsb.ToString();
-        }
-
-        /// <summary>Determines whether a value should be emitted directly rather than expanded into a query map.</summary>
-        /// <param name="value">The value to inspect.</param>
-        /// <returns><see langword="true"/> if the value is a simple/formattable type; otherwise <see langword="false"/>.</returns>
-        [RequiresUnreferencedCode("Refit's reflection-based request building is not trim-safe; use the Refit source generator for trimmed/AOT apps.")]
-        private static bool DoNotConvertToQueryMap(object? value)
-        {
-            if (value is null)
-            {
-                return false;
-            }
-
-            var type = value.GetType();
-
-            // Bail out early & match string
-            if (ShouldReturn(type))
-            {
-                return true;
-            }
-
-            if (value is not IEnumerable)
-            {
-                return false;
-            }
-
-            // Get the element type for enumerables
-            var ienu = typeof(IEnumerable<>);
-
-            // We don't want to enumerate to get the type, so we'll just look for IEnumerable<T>
-            Type? intType = null;
-            var interfaces = type.GetInterfaces();
-            for (var i = 0; i < interfaces.Length; i++)
-            {
-                var interfaceType = interfaces[i];
-                if (interfaceType.GetTypeInfo().IsGenericType
-                    && interfaceType.GetGenericTypeDefinition() == ienu)
-                {
-                    intType = interfaceType;
-                    break;
-                }
-            }
-
-            if (intType is null)
-            {
-                return false;
-            }
-
-            type = intType.GetGenericArguments()[0];
-            return ShouldReturn(type);
-
-            // Check if type is a simple string or IFormattable type, check underlying type if Nullable<T>
-            static bool ShouldReturn(Type type) =>
-                Nullable.GetUnderlyingType(type) is { } underlyingType
-                    ? ShouldReturn(underlyingType)
-                    : type == typeof(string)
-                      || type == typeof(bool)
-                      || type == typeof(char)
-                      || typeof(IFormattable).IsAssignableFrom(type)
-                      || type == typeof(Uri)
-                      || typeof(CultureInfo).IsAssignableFrom(type);
-        }
-
-        /// <summary>Sets or replaces a header on the request or its content, with CRLF-injection protection.</summary>
-        /// <param name="request">The request to modify.</param>
-        /// <param name="name">The header name.</param>
-        /// <param name="value">The header value, or null to only remove the header.</param>
-        private static void SetHeader(HttpRequestMessage request, string name, string? value)
-        {
-            // Clear any existing version of this header that might be set, because
-            // we want to allow removal/redefinition of headers.
-            // We also don't want to double up content headers which may have been
-            // set for us automatically.
-            // NB: We have to enumerate the header names to check existence because
-            // Contains throws if it's the wrong header type for the collection.
-            // HTTP header names are case-insensitive, so compare them that way; otherwise a
-            // differently cased header (e.g. "Content-type" vs "Content-Type") is not removed
-            // and ends up duplicated.
-            if (ContainsHeader(request.Headers, name))
-            {
-                request.Headers.Remove(name);
-            }
-
-            if (request.Content is not null && ContainsHeader(request.Content.Headers, name))
-            {
-                request.Content.Headers.Remove(name);
-            }
-
-            if (value is null)
-            {
-                return;
-            }
-
-            // CRLF injection protection
-            name = EnsureSafe(name);
-            value = EnsureSafe(value);
-
-            var added = request.Headers.TryAddWithoutValidation(name, value);
-
-            // Don't even bother trying to add the header as a content header
-            // if we just added it to the other collection.
-            if (added || request.Content is null)
-            {
-                return;
-            }
-
-            request.Content.Headers.TryAddWithoutValidation(name, value);
         }
 
         /// <summary>Strips CR and LF characters from a header value to prevent header injection.</summary>
@@ -425,7 +317,7 @@ namespace Refit
         [RequiresUnreferencedCode("Refit's reflection-based request building is not trim-safe; use the Refit source generator for trimmed/AOT apps.")]
         [RequiresDynamicCode("Refit's reflection-based request building requires runtime code generation; use the Refit source generator for AOT apps.")]
         private List<KeyValuePair<string, object?>> BuildQueryMap(
-            object? @object,
+            object @object,
             string? delimiter = null,
             RestMethodParameterInfo? parameterInfo = null,
             CollectionFormat? collectionFormat = null)
@@ -436,11 +328,6 @@ namespace Refit
             }
 
             var kvps = new List<KeyValuePair<string, object?>>();
-
-            if (@object is null)
-            {
-                return kvps;
-            }
 
             var props = GetCachedQueryProperties(@object.GetType());
             for (var i = 0; i < props.Length; i++)
@@ -529,14 +416,13 @@ namespace Refit
             var aliasAttribute = propertyInfo.GetCustomAttribute<AliasAsAttribute>();
             var key = aliasAttribute?.Name ?? _settings.UrlParameterKeyFormatter.Format(propertyInfo.Name);
 
-            // Look to see if the property has a Query attribute, and if so, format it accordingly
             var queryAttribute = propertyInfo.GetCustomAttribute<QueryAttribute>();
-            if (queryAttribute is { Format: not null })
+            if (!TryFormatQueryPropertyValue(queryAttribute, obj, out var formattedObj))
             {
-                obj = _settings.FormUrlEncodedParameterFormatter.Format(
-                    obj,
-                    queryAttribute.Format);
+                return;
             }
+
+            obj = formattedObj;
 
             // If obj is IEnumerable - format it accounting for Query attribute and CollectionFormat
             if (obj is not string and IEnumerable ienu and not IDictionary)
@@ -552,6 +438,38 @@ namespace Refit
             }
 
             AppendNestedQueryMap(obj, key, kvps, delimiter, collectionFormat);
+        }
+
+        /// <summary>Applies a property-level query format, if present.</summary>
+        /// <typeparam name="T">The value type.</typeparam>
+        /// <param name="queryAttribute">The query attribute, if any.</param>
+        /// <param name="value">The value to format.</param>
+        /// <param name="formattedValue">Receives the formatted value.</param>
+        /// <returns><see langword="true"/> when a non-null value remains.</returns>
+        [RequiresUnreferencedCode("Refit's reflection-based request building is not trim-safe; use the Refit source generator for trimmed/AOT apps.")]
+        [RequiresDynamicCode("Refit's reflection-based request building requires runtime code generation; use the Refit source generator for AOT apps.")]
+        private bool TryFormatQueryPropertyValue<T>(
+            QueryAttribute? queryAttribute,
+            T value,
+            [NotNullWhen(true)] out object? formattedValue)
+        {
+            if (queryAttribute is not { Format: not null })
+            {
+                formattedValue = value!;
+                return true;
+            }
+
+            var formatted = _settings.FormUrlEncodedParameterFormatter.Format(
+                value,
+                queryAttribute.Format);
+            if (formatted is null)
+            {
+                formattedValue = null;
+                return false;
+            }
+
+            formattedValue = formatted;
+            return true;
         }
 
         /// <summary>Appends each formatted element of an enumerable property to the query map under one key.</summary>
@@ -591,7 +509,7 @@ namespace Refit
         [RequiresUnreferencedCode("Refit's reflection-based request building is not trim-safe; use the Refit source generator for trimmed/AOT apps.")]
         [RequiresDynamicCode("Refit's reflection-based request building requires runtime code generation; use the Refit source generator for AOT apps.")]
         private void AppendNestedQueryMap(
-            object? obj,
+            object obj,
             string key,
             List<KeyValuePair<string, object?>> kvps,
             string? delimiter,
