+++ src/tests/Refit.Tests/CustomValueType.cs
@@ -2,10 +2,11 @@
 // ReactiveUI and Contributors licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-using System.Diagnostics.CodeAnalysis;
-
 namespace Refit.Tests;
 
 /// <summary>An empty custom type used to verify Refit's nullable handling of custom types.</summary>
-[SuppressMessage("RoslynCommonAnalyzers", "SST1436", Justification = "Intentional empty fixture type used to verify nullable handling in Refit tests.")]
+[System.Diagnostics.CodeAnalysis.SuppressMessage(
+    "RoslynCommonAnalyzers",
+    "SST1436:Add members to a type or remove it",
+    Justification = "Intentional empty fixture type used to verify nullable custom-type handling without changing the public shape.")]
 public sealed class CustomValueType;
