+++ src/tests/Refit.Tests/RestServiceIntegrationTests.Inheritance.cs
@@ -243,9 +243,12 @@ public partial class RestServiceIntegrationTests
                  "args": {"search.Addr.Street": "HomeStreet 99","search.Addr.Zip": "9999","search.FirstName": "John","search.LastName": "Rambo"}}
                 """);
 
-        var myParams = new MyComplexQueryParams { FirstName = "John", LastName = "Rambo" };
-        myParams.Address.Postcode = 9999;
-        myParams.Address.Street = "HomeStreet 99";
+        var myParams = new MyComplexQueryParams
+        {
+            FirstName = "John",
+            LastName = "Rambo",
+            Address = new() { Postcode = 9999, Street = "HomeStreet 99" },
+        };
 
         var fixture = RestService.For<IHttpBinApi<HttpBinGet, MyComplexQueryParams, int>>(
             "https://httpbin.org/get",
